requirements.txt
# AWS CDK Core
aws-cdk-lib==2.132.0
constructs==10.3.0

# AWS SDK
boto3==1.34.70
botocore==1.34.70
urllib3==1.26.16

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.1

# Development
pytest==8.0.2
black==24.2.0